fix(backend): resolve service registry test failures and mock leakage
